+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/registry.ts
@@ -14,7 +14,8 @@ import { ToggleClearFilterBarQuickAction } from './lr-toggle-clear-filter-bar';
 import { ChangeTableColumnsQuickAction } from './change-table-columns';
 import { AddHeaderFieldQuickAction } from '../common/op-add-header-field';
 import { AddCustomSectionQuickAction } from '../common/op-add-custom-section';
-import { AddNewAnnotationFile } from '../common/add-new-annotation-file';
+import { AddTableActionQuickAction } from '../fe-v2/create-table-action';
+import { AddPageActionQuickAction } from '../fe-v2/create-page-action';import { AddNewAnnotationFile } from '../common/add-new-annotation-file';
 
 type PageName = 'listReport' | 'objectPage';
 
@@ -42,7 +43,9 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                         ToggleClearFilterBarQuickAction,
                         AddControllerToPageQuickAction,
                         ChangeTableColumnsQuickAction,
-                        AddNewAnnotationFile
+                        AddNewAnnotationFile,
+                        AddTableActionQuickAction,
+                        AddPageActionQuickAction
                     ],
                     view,
                     key: name + index
@@ -55,7 +58,9 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                         ChangeTableColumnsQuickAction,
                         AddHeaderFieldQuickAction,
                         AddCustomSectionQuickAction,
-                        AddNewAnnotationFile
+                        AddNewAnnotationFile,
+                        AddTableActionQuickAction,
+                        AddPageActionQuickAction
                     ],
                     view,
                     key: name + index